requirements.txt
aiogram==3.2.0
fastapi==0.110.0
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.19
python-dotenv==1.0.0
aiohttp>=3.9.0
pydantic>=2.5
hdbcli
requests